requirements.txt
python-telegram-bot==13.13
bs4
requests
